app/requirements.txt
flask
boto3
google-cloud-storage

